mutualize live camera method retrieval
